Merge branch '199430_min_monte_conciliacion_f2' of https://github.com/MontePiedadMx/MS-Pagos-Prendarios into 199430_min_monte_conciliacion_f2

# Conflicts:
#	src/main/java/mx/com/nmp/pagos/mimonte/controllers/conciliacion/ConciliacionController.java
#	src/main/java/mx/com/nmp/pagos/mimonte/controllers/conciliacion/LayoutsController.java
#	src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/ComisionesRepository.java
#	src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/MovimientoConciliacionRepository.java
#	src/main/java/mx/com/nmp/pagos/mimonte/services/conciliacion/ComisionesService.java
#	src/main/java/mx/com/nmp/pagos/mimonte/services/conciliacion/LayoutsService.java
#	src/main/java/mx/com/nmp/pagos/mimonte/services/impl/conciliacion/ConciliacionServiceImpl.java
#	src/main/java/mx/com/nmp/pagos/mimonte/services/impl/conciliacion/DevolucionesServiceImpl.java

+++ src/main/java/mx/com/nmp/pagos/mimonte/services/conciliacion/LayoutsService.java
@@ -5,6 +5,7 @@
 package mx.com.nmp.pagos.mimonte.services.conciliacion;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -14,6 +15,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -45,6 +48,7 @@ import mx.com.nmp.pagos.mimonte.model.conciliacion.MovimientoConciliacion;
 import mx.com.nmp.pagos.mimonte.model.conciliacion.MovimientoDevolucion;
 import mx.com.nmp.pagos.mimonte.model.conciliacion.MovimientoPago;
 import mx.com.nmp.pagos.mimonte.model.conciliacion.TipoLayoutEnum;
+import mx.com.nmp.pagos.mimonte.util.ConciliacionDataValidator;
 
 /**
  * * @name LayoutsService
@@ -100,6 +104,20 @@ public class LayoutsService {
 	@Autowired
 	private ConciliacionHelper conciliacionHelper;
 
+	/**
+	 * Validador generico de conciliacion
+	 */
+	@Autowired
+	private ConciliacionDataValidator conciliacionDataValidator;
+	
+	/**
+	 * Logs de la clase
+	 */
+	private static final Logger LOG = LoggerFactory.getLogger(LayoutsService.class);
+	
+	public LayoutsService() {
+		super();
+	}
 
 
 
@@ -138,18 +156,41 @@ public class LayoutsService {
 	 * @return
 	 */
 	public List<LayoutDTO> consultarLayouts(Long idConciliacion) {
-		List<Layout> layouts = new ArrayList<Layout>();
+		// Objetos necesarios
+		List<LayoutDTO> layoutDTOs = new ArrayList<>();
+		
 		try {
-			layouts = layoutsRepository.findByIdConciliacion(idConciliacion);
+			// Valida que la conciliacion exista
+			conciliacionDataValidator.validateFolioExists(idConciliacion);
+			
+			if (idConciliacion > 0L) {
+				LayoutDTO layoutDTO = consultarUnLayout(idConciliacion, TipoLayoutEnum.PAGOS);
+				if (layoutDTO != null) {
+					layoutDTOs.add(layoutDTO);
+				}
+				layoutDTO = consultarUnLayout(idConciliacion, TipoLayoutEnum.COMISIONES_MOV);
+				if (layoutDTO != null) {
+					layoutDTOs.add(layoutDTO);
+				}
+				layoutDTO = consultarUnLayout(idConciliacion, TipoLayoutEnum.COMISIONES_GENERALES);
+				if (layoutDTO != null) {
+					layoutDTOs.add(layoutDTO);
+				}
+				layoutDTO = consultarUnLayout(idConciliacion, TipoLayoutEnum.DEVOLUCIONES);
+				if (layoutDTO != null) {
+					layoutDTOs.add(layoutDTO);
+				}
+			}	
 		}
-		catch (Exception ex) {
-			ex.printStackTrace();
-			throw new ConciliacionException("Error al consultar los layouts", CodigoError.NMP_PMIMONTE_BUSINESS_011);
+		catch(ConciliacionException ex) {
+			LOG.debug(ConciliacionConstants.GENERIC_EXCEPTION_INITIAL_MESSAGE.concat("{}"), ex.getMessage());
+			throw ex;
 		}
-		if (layouts == null) {
-			throw new ConciliacionException("No existe un layouts para la conciliacion " + idConciliacion, CodigoError.NMP_PMIMONTE_BUSINESS_084);
+		catch(Exception ex) {
+			LOG.debug(ConciliacionConstants.GENERIC_EXCEPTION_INITIAL_MESSAGE.concat("{}"), ex.getMessage());
+			throw new ConciliacionException(CodigoError.NMP_PMIMONTE_BUSINESS_113.getDescripcion(), CodigoError.NMP_PMIMONTE_BUSINESS_113);
 		}
-		return LayoutsBuilder.buildLayoutsDTOFromLayouts(layouts);
+		return layoutDTOs;
 	}
 
 
@@ -190,26 +231,56 @@ public class LayoutsService {
 	 */
 	@Transactional
 	public boolean eliminarUnLayout(Long idConciliacion, Long idLayout) throws ConciliacionException {
-		boolean respuesta = true;
-		List<Layout> layoutList = layoutsRepository.checkFolioAndLayoutsRelationship(idConciliacion);
-		if (null == layoutList || layoutList.isEmpty())
-			throw new ConciliacionException(ConciliacionConstants.THERE_IS_NO_CONCILIACION_LAYOUT_RELATIONSHIP,
-					CodigoError.NMP_PMIMONTE_BUSINESS_084);
-		if (idConciliacion > 0L && idLayout > 0L) {
-			if (layoutsRepository.findById(idLayout).isPresent()) {
-				Layout layout = layoutsRepository.getOne(idLayout);// siLayoutPerteneceAConciliacion
-				List<Layout> layouts = layoutsRepository.findByTipo(layout.getTipo());// siLayoutEsElMismoTipo
-				for (Layout layout2 : layouts) {
-					if (layout2.getIdConciliacion() == idConciliacion) {
-						layoutsRepository.eliminarUnLayoutLineas(layout2.getId());
-						layoutsRepository.eliminarUnLayoutHeader(layout2.getId());
-						layoutsRepository.deleteById(layout2.getId());
-						respuesta = false;
+		// Objetos necesarios
+		Boolean flag = null;
+		boolean respuesta = false;
+		
+		// LOGICA DE ELIMINACION
+		try {
+			// Valida que la conciliacion exista
+			conciliacionDataValidator.validateFolioExists(idConciliacion);
+			
+			// Valida que el id de layout exista
+			flag = ((BigInteger)layoutsRepository.checkIfIdExist(idLayout)).compareTo(BigInteger.ONE) == 0;
+			if(!flag)
+				throw new ConciliacionException(CodigoError.NMP_PMIMONTE_BUSINESS_085.getDescripcion(), CodigoError.NMP_PMIMONTE_BUSINESS_085);
+			
+			// Valida que la conciliacion este ligada al id de layout especificado
+			flag = null;
+			flag = ((BigInteger)layoutsRepository.checkIfFolioIdRelationshipExist(idConciliacion.intValue(), idLayout)).compareTo(BigInteger.ONE) == 0;
+			if(!flag)
+				throw new ConciliacionException(CodigoError.NMP_PMIMONTE_BUSINESS_115.getDescripcion(), CodigoError.NMP_PMIMONTE_BUSINESS_115);
+			
+			// Logica de eliminacion
+			respuesta = true;
+			List<Layout> layoutList = layoutsRepository.checkFolioAndLayoutsRelationship(idConciliacion);
+			if (null == layoutList || layoutList.isEmpty())
+				throw new ConciliacionException(ConciliacionConstants.THERE_IS_NO_CONCILIACION_LAYOUT_RELATIONSHIP,
+						CodigoError.NMP_PMIMONTE_BUSINESS_084);
+			if (idConciliacion > 0L && idLayout > 0L) {
+				if (layoutsRepository.findById(idLayout).isPresent()) {
+					Layout layout = layoutsRepository.getOne(idLayout);// siLayoutPerteneceAConciliacion
+					List<Layout> layouts = layoutsRepository.findByTipo(layout.getTipo());// siLayoutEsElMismoTipo
+					for (Layout layout2 : layouts) {
+						if (layout2.getIdConciliacion() == idConciliacion) {
+							layoutsRepository.eliminarUnLayoutLineas(layout2.getId());
+							layoutsRepository.eliminarUnLayoutHeader(layout2.getId());
+							layoutsRepository.deleteById(layout2.getId());
+							respuesta = false;
+						}
 					}
-				}
-			} else
-				throw new ConciliacionException(ConciliacionConstants.LAYOUT_ID_DOESNT_EXIST,
-						CodigoError.NMP_PMIMONTE_BUSINESS_085);
+				} else
+					throw new ConciliacionException(ConciliacionConstants.LAYOUT_ID_DOESNT_EXIST,
+							CodigoError.NMP_PMIMONTE_BUSINESS_085);
+			}	
+		}
+		catch(ConciliacionException ex) {
+			LOG.debug(ConciliacionConstants.GENERIC_EXCEPTION_INITIAL_MESSAGE.concat("{}"), ex.getMessage());
+			throw ex;
+		}
+		catch(Exception ex) {
+			LOG.debug(ConciliacionConstants.GENERIC_EXCEPTION_INITIAL_MESSAGE.concat("{}"), ex.getMessage());
+			throw new ConciliacionException(CodigoError.NMP_PMIMONTE_BUSINESS_114.getDescripcion(), CodigoError.NMP_PMIMONTE_BUSINESS_114);
 		}
 		return respuesta;
 	}
